Changed aliases to better support proper commit messages
I have been lazy in the way that I handle commit messages, this change
was done to support this new format. Not every commit message needs a
full text write up, and so I have aliased dotcm to do that easily. For
larger changes that would benefit from a more detailed write up I have
created dotc to do so.