Added i3 files incase xorg is needed
There are still a number of applications that do not work correctly
under xwayland (obs-studio for example), ans so I have started porting
my install over to i3. The idea is to make things as simular and
seamless as possible, while allowing me to use i3 and xorg when
required.